+++ src/services/TimeService.ts
@@ -2,8 +2,8 @@ import {injectable} from 'inversify';
 
 @injectable()
 class TimeService {
-  apply(): number {
-    return Math.floor(Date.now() / 1000);
+  apply(offset = 0): number {
+    return Math.floor(Date.now() / 1000) - offset;
   }
 }