+++ src/types/SignedBlock.ts
@@ -3,9 +3,9 @@ export interface KeyValues {
 }
 
 export interface SignedBlock {
+  blockHeight: number;
+  dataTimestamp: number;
   fcd: KeyValues;
   leaves: KeyValues;
-  blockHeight: number;
   signature: string;
-  dataTimestamp: number;
 }