+++ src/types/ValidatorsResponses.ts
@@ -1,0 +1,7 @@
+import {BigNumber} from 'ethers';
+
+export interface ValidatorsResponses {
+  signatures: string[];
+  discrepanciesKeys: Set<string>;
+  powers: BigNumber;
+}