+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Added
 - added new event `PriceDiscrepancy` reported to NewRelic
+- self-adjusting consensus, that can ignore discrepancy pairs
 
 ## [1.0.1] - 2021-04-16